Add version property for logback-classic
We have been inheriting the logback.version property from the
kiwi-parent POM, but this has been removed from kiwi-parent.
The logback-classic and logback-core dependencies are already
defined in this BOM, but this commit introduces separate version
properties so that, if one ever has a patch version while other
does not, we can accomodate by changing one or the other of
logback-classic.verion or logback.core.version. Of course, this does
mean if they both change, we have to change both of them, but this
doesn't happen too often and they are right next to each other.

Closes #167