Revert logback-classic and logback-core to 1.5.12
This is the highest version of these libraries that we can
use with Dropwizard 4.x (and even the alpha versions of 5.x).
We also must remain on logback-access 1.4.14 or else Dropwizard
code can't work because it depends on the 1.4.x logback-access,
which has a different package than the newer logback-access 2.0.x
versions.

Dropwizard needs to update their usage of Logback classic, core,
and access for us to move beyond these versions. :-(

Fixes #1239